requirements.txt
google-generativeai
playwright
python-dotenv
numpy
pandas
scikit-learn
requests
gguf